Move filterAndReject to /src/lib
